created modxo_modules, refactored modxo source to encapsulate modules.
