+++ modxo/config/CMakeLists.txt
@@ -1,0 +1,1 @@
+target_sources(modxo PRIVATE ${CMAKE_CURRENT_LIST_DIR}/config_nvm.c )
