+++ modxo/data_store/CMakeLists.txt
@@ -1,0 +1,1 @@
+target_sources(modxo PRIVATE data_store.c)
