+++ modxo/include/modxo/lpc_log.h
