+++ modxo/include/modxo/lpc_regs.h
