+++ modxo/lpc/lpc_comm.in
