+++ modxo/lpc/lpc_interface.c
@@ -29,9 +29,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <stdio.h>
 #include "pico/stdlib.h"
 #include "hardware/pio.h"
-#include "hardware/dma.h"
 #include "hardware/irq.h"
-#include "pico/multicore.h"
 #include "hardware/structs/padsbank0.h"
 #include "hardware/structs/bus_ctrl.h"
 #include <modxo/lpc_interface.h>
