+++ modxo/lpc/lpc_log.c
