+++ modxo/lpc/lpc_regs.c
