+++ modxo_modules/flashrom/include/flashrom.h
@@ -10,12 +10,5 @@ Copyright (c) 2024, Shalx <Alejandro L. Huitron shalxmva@gmail.com>
 #include <modxo.h>
 #include <stdint.h>
 
-#define MODXO_BANK_BOOTLOADER 0x01
-
-void flashrom_set_mmc(uint8_t);
-uint8_t flashrom_get_mmc(void);
-void flashrom_erase_sector(uint8_t sectorn);
-void flashrom_program_sector(uint8_t sectorn);
-
 extern MODXO_TASK flashrom_hdlr;
 #endif
