--- src/flashrom/CMakeLists.txt
@@ -1,2 +1,0 @@
-target_sources(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_LIST_DIR}/flashrom.c )
-target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR}/include)
