--- src/legacy_display/CMakeLists.txt
@@ -1,2 +1,0 @@
-target_sources(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_LIST_DIR}/legacy_display.c )
-target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR}/include)
