--- src/modxo/CMakeLists.txt
@@ -1,7 +1,0 @@
-
-target_sources(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR}/modxo_ports.c )
-target_sources(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR}/modxo.c )
-target_sources(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR}/modxo_queue.c )
-target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR}/include)
-add_subdirectory(lpc)
-add_subdirectory(config)
