--- src/modxo/config/CMakeLists.txt
@@ -1,1 +1,0 @@
-target_sources(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_LIST_DIR}/config_nvm.c )
